Logged into Nexus Mods from macOS
