+++ NexusMods.App.sln
@@ -269,11 +269,16 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.Abstractions.Resources.Resilience", "src\Abstractions\NexusMods.Abstractions.Resources.Resilience\NexusMods.Abstractions.Resources.Resilience.csproj", "{04219A58-C99C-4C3B-A477-5E4B29D1F275}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.Abstractions.UI", "src\Abstractions\NexusMods.Abstractions.UI\NexusMods.Abstractions.UI.csproj", "{A7417BB1-7E2C-413A-A999-546F021F683B}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.DataModel.SchemaVersions", "src\NexusMods.DataModel.SchemaVersions\NexusMods.DataModel.SchemaVersions.csproj", "{79E13AD1-187B-42F7-BDC3-EF8ABA308973}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.DataModel.SchemaVersions.Tests", "tests\NexusMods.DataModel.SchemaVersions.Tests\NexusMods.DataModel.SchemaVersions.Tests.csproj", "{A5A2932D-B3EF-480B-BEBC-793F6FC90EDE}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.Games.MountAndBlade2Bannerlord", "src\Games\NexusMods.Games.MountAndBlade2Bannerlord\NexusMods.Games.MountAndBlade2Bannerlord.csproj", "{8D7E82BB-2F8D-455A-AF12-C486D9EC3B77}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.App.MacOS", "src\NexusMods.App.MacOS\NexusMods.App.MacOS.csproj", "{8375D256-3C4F-437E-911E-5EC101422A9A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.App.Common", "src\NexusMods.App.Common\NexusMods.App.Common.csproj", "{28E7A782-DCC9-4769-A209-8D4601BCEFEC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -721,6 +726,14 @@ Global
 		{8D7E82BB-2F8D-455A-AF12-C486D9EC3B77}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8D7E82BB-2F8D-455A-AF12-C486D9EC3B77}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8D7E82BB-2F8D-455A-AF12-C486D9EC3B77}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8375D256-3C4F-437E-911E-5EC101422A9A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8375D256-3C4F-437E-911E-5EC101422A9A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8375D256-3C4F-437E-911E-5EC101422A9A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8375D256-3C4F-437E-911E-5EC101422A9A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{28E7A782-DCC9-4769-A209-8D4601BCEFEC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{28E7A782-DCC9-4769-A209-8D4601BCEFEC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{28E7A782-DCC9-4769-A209-8D4601BCEFEC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{28E7A782-DCC9-4769-A209-8D4601BCEFEC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -849,6 +862,8 @@ Global
 		{79E13AD1-187B-42F7-BDC3-EF8ABA308973} = {E7BAE287-D505-4D6D-A090-665A64309B2D}
 		{A5A2932D-B3EF-480B-BEBC-793F6FC90EDE} = {52AF9D62-7D5B-4AD0-BA12-86F2AA67428B}
 		{8D7E82BB-2F8D-455A-AF12-C486D9EC3B77} = {70D38D24-79AE-4600-8E83-17F3C11BA81F}
+		{8375D256-3C4F-437E-911E-5EC101422A9A} = {E7BAE287-D505-4D6D-A090-665A64309B2D}
+		{28E7A782-DCC9-4769-A209-8D4601BCEFEC} = {E7BAE287-D505-4D6D-A090-665A64309B2D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9F9F8352-34DD-42C0-8564-EE9AF34A3501}
