+++ src/NexusMods.App.Common/App.axaml.cs
@@ -9,11 +9,12 @@ using NexusMods.Abstractions.Settings;
 using NexusMods.App.UI;
 using NexusMods.App.UI.Settings;
 using NexusMods.App.UI.Windows;
+using NexusMods.CLI.Types;
 using ReactiveUI;
 using Splat;
 using Splat.Microsoft.Extensions.Logging;
 
-namespace NexusMods.App;
+namespace NexusMods.App.Common;
 
 public class App : Application
 {
@@ -26,6 +27,7 @@ public class App : Application
         _provider = provider;
         _settingsManager = provider.GetRequiredService<ISettingsManager>();
         Name = "Nexus Mods";
+        
     }
 
     public App()
@@ -33,6 +35,8 @@ public class App : Application
         throw new UnreachableException("We don't use the Runtime Loader, so this should never be called.");
     }
 
+    
+    
     public override void Initialize()
     {
         AvaloniaXamlLoader.Load(this);
@@ -58,6 +62,26 @@ public class App : Application
             var reactiveWindow = _provider.GetRequiredService<MainWindow>();
             reactiveWindow.ViewModel = _provider.GetRequiredService<MainWindowViewModel>();
             desktop.MainWindow = reactiveWindow;
+
+            var activatableLifetime = TryGetFeature(typeof(IActivatableLifetime));
+            if (activatableLifetime is IActivatableLifetime activatable)
+            {
+                activatable.Activated += async (_, args) =>
+                {
+                    switch (args.Kind)
+                    {
+                        case ActivationKind.OpenUri:
+                            var protocol = (ProtocolActivatedEventArgs)args;
+                            logger.LogInformation("Got Url {Url}", protocol.Uri.ToString());
+                            var handler = _provider.GetRequiredService<IIpcProtocolHandler>();
+                            await handler.Handle(protocol.Uri.ToString(), CancellationToken.None);
+                            break;
+                        default:
+                            break;
+                    }
+                };
+            }
+            
         }
 
         base.OnFrameworkInitializationCompleted();
