+++ src/NexusMods.App.MacOS/Main.cs
@@ -1,0 +1,6 @@
+
+
+// This is the main entry point of the application.
+
+NSApplication.Init ();
+NexusMods.App.Common.Program.Main(args);
