+++ src/NexusMods.App.MacOS/ViewController.cs
@@ -1,0 +1,27 @@
+using ObjCRuntime;
+
+namespace NexusMods.App.Mac;
+
+public partial class ViewController : NSViewController {
+	protected ViewController (NativeHandle handle) : base (handle)
+	{
+		// This constructor is required if the view controller is loaded from a xib or a storyboard.
+		// Do not put any initialization here, use ViewDidLoad instead.
+	}
+
+	public override void ViewDidLoad ()
+	{
+		base.ViewDidLoad ();
+
+		// Do any additional setup after loading the view.
+	}
+
+	public override NSObject RepresentedObject {
+		get => base.RepresentedObject;
+		set {
+			base.RepresentedObject = value;
+
+			// Update the view, if already loaded.
+		}
+	}
+}
